// src/types/generated/OAuthFlow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * How the user finishes a login that [`OAuthStartInfo`] just started.
 */
export type OAuthFlow = "local-callback" | "remote-poll" | { "scheme-paste": { scheme_prefix: string, } } | "immediate";